+++ tests/login/test_main.py
@@ -25,7 +25,7 @@ def test_login_logout(
     expect(page.get_by_test_id("start-page-body")).to_be_visible()
     page.screenshot(path="tests_login_test_main-test_login_logout-after-login.png")
 
-    page.get_by_test_id("user-menu").click()
+    expect(page.get_by_test_id("user-menu")).to_be_visible()
     expect(page.get_by_test_id("logout-button")).to_be_visible()
     page.get_by_test_id("logout-button").click()
     page.wait_for_url(base_url)